Move file system types to shared module
