+++ front/lib/api/file_system/index.ts
@@ -4,14 +4,14 @@ export {
   parseScopedPrefix,
   sanitizeFileSystemName,
 } from "@app/lib/api/file_system/dust_file_system";
+export type { FileSystemEntry } from "@app/types/api/file_system/types";
 export type {
   DustFileSystemErrorCode,
   FileSystemMount,
-} from "@app/lib/api/file_system/types";
+} from "@app/types/file_system";
 export {
   LEGACY_PREFIX_CONVERSATION,
   LEGACY_PREFIX_PROJECT,
   SCOPED_PREFIX_CONVERSATION,
   SCOPED_PREFIX_POD,
-} from "@app/lib/api/file_system/types";
-export type { FileSystemEntry } from "@app/types/api/file_system/types";
+} from "@app/types/file_system";
