+++ front/lib/api/files/action_output_fs/index.ts
@@ -2,10 +2,6 @@ import { FILE_OFFLOAD_TEXT_SIZE_BYTES } from "@app/lib/actions/action_output_lim
 import { isResourceContentWithText } from "@app/lib/actions/mcp_internal_actions/output_schemas";
 import type { ToolRunContext } from "@app/lib/actions/types";
 import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
-import {
-  conversationScopedPath,
-  podScopedPath,
-} from "@app/lib/api/file_system/types";
 import { makeFileName } from "@app/lib/api/files/action_output_fs/naming";
 import {
   resolveResourceOutput,
@@ -15,6 +11,7 @@ import { TOOL_OUTPUTS_FOLDER_NAME } from "@app/lib/api/files/mount_path";
 import type { Authenticator } from "@app/lib/auth";
 import type { SpaceResource } from "@app/lib/resources/space_resource";
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
+import { conversationScopedPath, podScopedPath } from "@app/types/file_system";
 import type { AllSupportedFileContentType } from "@app/types/files";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
