+++ front/lib/api/files/file_system_ops.ts
@@ -5,16 +5,16 @@
  */
 
 import type { DustFileSystem } from "@app/lib/api/file_system";
-import {
-  DustFileSystemError,
-  SCOPED_PREFIX_CONVERSATION,
-  SCOPED_PREFIX_POD,
-} from "@app/lib/api/file_system/types";
 import { decodeBuffer } from "@app/lib/api/files/utils";
 import type { Authenticator } from "@app/lib/auth";
 import { FileResource } from "@app/lib/resources/file_resource";
 import logger from "@app/logger/logger";
 import type { FileSystemEntry } from "@app/types/api/file_system/types";
+import {
+  DustFileSystemError,
+  SCOPED_PREFIX_CONVERSATION,
+  SCOPED_PREFIX_POD,
+} from "@app/types/file_system";
 import type { FileUseCase, FileUseCaseMetadata } from "@app/types/files";
 import {
   contentTypeFromFileName,
