+++ front/lib/api/projects/context.ts
@@ -6,10 +6,6 @@ import { getContentFragmentBlob } from "@app/lib/api/assistant/conversation/cont
 import { getContentNodesForDataSourceView } from "@app/lib/api/data_source_view";
 import { DustFileSystem } from "@app/lib/api/file_system";
 import {
-  DustFileSystemError,
-  podScopedPath,
-} from "@app/lib/api/file_system/types";
-import {
   deleteGCSMountFile,
   moveFile,
   renameGCSMountDirectory,
@@ -40,6 +36,7 @@ import type { ConversationAttachmentType } from "@app/types/api/assistant/conver
 import type { FileSystemDirectoryEntry } from "@app/types/api/file_system/types";
 import type { ContentNodeType } from "@app/types/core/content_node";
 import type { ConnectorProvider } from "@app/types/data_source";
+import { DustFileSystemError, podScopedPath } from "@app/types/file_system";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { removeNulls } from "@app/types/shared/utils/general";
