+++ front/lib/api/user_memory.ts
@@ -1,8 +1,8 @@
 import type { DustFileSystemError } from "@app/lib/api/file_system/dust_file_system";
 import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
-import { userScopedPath } from "@app/lib/api/file_system/types";
 import type { Authenticator } from "@app/lib/auth";
 import { MAX_USER_MEMORY_CONTENT_LENGTH } from "@app/types/api/me/memory";
+import { userScopedPath } from "@app/types/file_system";
 import type { Result } from "@app/types/shared/result";
 import { Ok } from "@app/types/shared/result";
